Code splitting into seperate components from Main component
